Split rqh and portal api (#18)
* feat: Add react-query module to libs/index.ts

This commit adds the export statement for the react-query module in the index.ts file of the libs directory. This allows other parts of the codebase to import and use the react-query module.

* chore: Update imports for sign-in functionality

The code changes update the imports for sign-in functionality in the SignIn component and its related Form component. The AuthService import is replaced with the signInWithGoogle and signInWithEmailAndPassword imports from the services module. This improves the organization and clarity of the code.

* chore: Update environment variable names for API base URLs

This commit updates the environment variable names for the API base URLs in the `.env.example` file. The `NEXT_PUBLIC_API_BASE_URL` variable is replaced with `NEXT_PUBLIC_RQH_API_BASE_URL` and `NEXT_PUBLIC_PORTAL_API_BASE_URL`. This change aligns with the recent updates made to the GitHub workflows and ensures consistency across the codebase.

* chore: Update environment variable names for API base URLs